backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.8.1
openai==1.3.8
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.1
httpx==0.25.2
beautifulsoup4==4.12.2
markdown==3.5.1
python-wordpress-xmlrpc==2.3 